Added airport deduplication and optimizations to airport delete processor. Increased distance for VOR and NDB deduplication. Printing duplicated airports into log before deleting. Now appending scenery path and bgl filenames in airport table. Added result set printing to sql script. Added number of starts to airport for proper delete handling. Added table airport_file to collect all sceneries containing an airport. Added compiler option for deduplication.

+++ src/fs/bgl/record.h
@@ -75,7 +75,7 @@ public:
    */
   int getSize() const
   {
-    return size;
+    return static_cast<int>(size);
   }
 
   /* Get the record id casted into an enum */
@@ -106,7 +106,7 @@ public:
 protected:
   friend QDebug operator<<(QDebug out, const atools::fs::bgl::Record& record);
 
-  int id, size;
+  unsigned int id, size;
   bool excluded = false;
 };
